code/requirements.txt
openai
pandas
python-dotenv
langchain
langchain-community
langchain-openai
chromadb
sentence-transformers